backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
firebase-admin==6.5.0
apscheduler==3.10.4
pydantic==2.9.0
python-dotenv==1.0.1
